Move archive out of plugincheck to its own package
